Extract poll chart calculations into a shared concern
