+++ app/controllers/api/v1/constituencies_controller.rb
@@ -4,6 +4,8 @@ module Api
     # by-election, the full list for a general election. Reference data, same
     # as parties: no auth, no phase gate.
     class ConstituenciesController < BaseController
+      include PollChartData
+
       def index
         render json: ConstituencySerializer.new(
           OnsConstituency.all.order(:name)
@@ -28,33 +30,6 @@ module Api
       end
 
       private
-
-      # Matches PollsHelper#poll_data_for: parties with no predicted votes are
-      # left off the chart entirely, and the rest run biggest first.
-      def chart_polls(all_polls)
-        all_polls.reject { |poll| poll.votes.to_i.zero? }
-                 .sort_by { |poll| -poll.votes }
-      end
-
-      # Same maths as Poll#signed_marginal_score, but computed here from the
-      # constituency's polls we already loaded, rather than each poll
-      # re-querying its constituency and siblings (no `inverse_of` on
-      # OnsConstituency#polls means that association isn't primed either
-      # way). Comparing against `all_polls` — not the zero-vote-filtered
-      # `polls` — matches the model method: a poll's margin is against every
-      # other party standing, not just the ones the chart draws.
-      #
-      # `others.max` can be nil (a constituency with a single poll, or every
-      # sibling having no recorded votes yet), and any sibling's `votes` can
-      # be nil (the column has no default) — both would blow up the model's
-      # `votes - other_poll_votes.max`, so we treat a missing comparison as
-      # "nothing to be ahead or behind of", i.e. 0.
-      def signed_marginal_scores(polls, all_polls)
-        polls.each_with_object({}) do |poll, scores|
-          others = all_polls.reject { |other| other.id == poll.id }.filter_map(&:votes)
-          scores[poll.id] = poll.votes - (others.max || 0)
-        end
-      end
     end
   end
 end
